Extract backend-independent SIR crate
