+++ crates/celox-sir/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name                  = "celox-sir"
+version.workspace     = true
+authors.workspace     = true
+repository.workspace  = true
+license.workspace     = true
+description           = "Backend-independent Simulator IR for Celox"
+edition.workspace     = true
+
+[dependencies]
+celox-analysis = { path = "../celox-analysis" }
+fxhash         = { workspace = true }
+num-bigint     = { workspace = true }
+num-traits     = { workspace = true }
+serde          = { workspace = true }
+
+[lints]
+workspace = true
