+++ crates/celox/src/ir.rs
@@ -2,11 +2,16 @@ use crate::{
     HashMap, HashSet,
     logic_tree::{LogicPath, SLTNodeArena, SymbolicStore},
 };
+pub(crate) use celox_sir::{
+    BasicBlock, BinaryOp, BlockId, DomainKind, ExecutionUnit, RegisterId, RegisterType, SIRBuilder,
+    SIRInstruction, SIROffset, SIRSwitchCase, SIRTerminator, SIRValue, SirMergeProvenance,
+    TriggerIdWithKind, UnaryOp, collect_exact_zero_registers, inline_single_predecessor_jumps,
+    merge_sir_eu_refs_with_provenance, merge_sir_eus,
+};
 use num_bigint::BigUint;
-use num_traits::Zero;
 use serde::{Deserialize, Serialize};
+use std::collections::BTreeSet;
 use std::fmt;
-use std::{collections::BTreeSet, fmt::Display};
 use veryl_analyzer::ir::{VarId, VarPath, Variable};
 
 /// Error returned by [`Program::get_addr`] when a path-based variable lookup fails.
@@ -20,15 +25,6 @@ pub enum AddrLookupError {
     AmbiguousPath { path: String },
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum DomainKind {
-    ClockPosedge,
-    ClockNegedge,
-    ResetAsyncHigh,
-    ResetAsyncLow,
-    Other,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum PortTypeKind {
     Clock,
@@ -39,12 +35,6 @@ pub enum PortTypeKind {
     Logic,
     Bit,
     Other,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct TriggerIdWithKind {
-    pub kind: DomainKind,
-    pub id: usize,
 }
 
 #[derive(Clone)]
@@ -523,14 +513,13 @@ where
     }
 }
 pub type VarAtom = VarAtomBase<VarId>;
-mod builder;
-pub(crate) use builder::SIRBuilder;
-pub(crate) mod cfg;
-pub(crate) mod verify;
+pub(crate) mod cfg {
+    pub(crate) use celox_sir::cfg::*;
+}
+pub(crate) mod verify {
+    pub(crate) use celox_sir::verify::*;
+}
 use veryl_parser::resource_table::StrId;
-/// Block identifier
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct BlockId(pub usize);
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(bound(
     serialize = "V: Serialize",
@@ -684,59 +673,6 @@ impl fmt::Debug for RelocationModule {
             .finish()
     }
 }
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(bound(serialize = "A: Serialize", deserialize = "A: Deserialize<'de>"))]
-pub struct ExecutionUnit<A> {
-    pub entry_block_id: BlockId,
-    pub blocks: HashMap<BlockId, BasicBlock<A>>,
-    pub register_map: HashMap<RegisterId, RegisterType>,
-}
-
-impl<A: Display> Display for ExecutionUnit<A> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        writeln!(f, "ExecutionUnit {{")?;
-        writeln!(f, "  entry: b{}", self.entry_block_id.0)?;
-        writeln!(f, "  registers: {{")?;
-        let mut reg_ids: Vec<_> = self.register_map.keys().collect();
-        reg_ids.sort();
-        for id in reg_ids {
-            let ty = &self.register_map[id];
-            match ty {
-                RegisterType::Logic { width } => {
-                    writeln!(f, "    r{}: logic<{}>", id.0, width)?;
-                }
-                RegisterType::Bit { width, signed } => {
-                    let s = if *signed { "signed " } else { "" };
-                    writeln!(f, "    r{}: {}bit<{}>", id.0, s, width)?;
-                }
-            }
-        }
-        writeln!(f, "  }}")?;
-        let block_order = crate::cfg_order::dominance_order(
-            self.entry_block_id,
-            self.blocks.keys().copied(),
-            |id| match &self.blocks[&id].terminator {
-                SIRTerminator::Jump(target, _) => vec![*target],
-                SIRTerminator::Branch {
-                    true_block,
-                    false_block,
-                    ..
-                } => vec![true_block.0, false_block.0],
-                SIRTerminator::Switch { cases, default, .. } => cases
-                    .iter()
-                    .map(|case| case.target)
-                    .chain(std::iter::once(*default))
-                    .collect(),
-                SIRTerminator::Return | SIRTerminator::Error(_) => Vec::new(),
-            },
-        );
-        for id in block_order {
-            let block = &self.blocks[&id];
-            writeln!(f, "{}", block)?;
-        }
-        writeln!(f, "}}")
-    }
-}
 #[derive(Clone)]
 pub struct SimModule {
     pub name: StrId,
@@ -800,1342 +736,6 @@ impl SimModule {
     }
 }
 
-/// Merge multiple SIR ExecutionUnits into a single EU.
-/// Each EU's Return becomes a Jump to the next EU's entry block.
-/// RegisterIds and BlockIds are renumbered to avoid conflicts.
-/// Returns (merged_eu, eu_entry_block_ids) where eu_entry_block_ids[i] is the
-/// BlockId of the i-th EU's entry block in the merged EU (for i > 0).
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-pub fn merge_sir_eus<A: Clone>(units: &[ExecutionUnit<A>]) -> (ExecutionUnit<A>, Vec<BlockId>) {
-    let units = units.iter().collect::<Vec<_>>();
-    merge_sir_eu_refs(&units)
-}
-
-/// Exact source-unit provenance for a merged SIR function.
-///
-/// Unlike a list of boundary block IDs, this remains valid when an input
-/// execution unit has a nonzero entry ID or sparse block IDs.
-#[derive(Debug, Clone)]
-pub(crate) struct SirMergeProvenance {
-    pub unit_entries: Vec<BlockId>,
-    #[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-    pub block_units: crate::HashMap<BlockId, usize>,
-}
-
-/// Reference-based variant of [`merge_sir_eus`] used when one compilation
-/// unit is assembled from multiple Program-owned EU slices.
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-pub(crate) fn merge_sir_eu_refs<A: Clone>(
-    units: &[&ExecutionUnit<A>],
-) -> (ExecutionUnit<A>, Vec<BlockId>) {
-    let (merged, provenance) = merge_sir_eu_refs_with_provenance(units);
-    (merged, provenance.unit_entries[1..].to_vec())
-}
-
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-pub(crate) fn merge_sir_eu_refs_with_provenance<A: Clone>(
-    units: &[&ExecutionUnit<A>],
-) -> (ExecutionUnit<A>, SirMergeProvenance) {
-    assert!(!units.is_empty(), "cannot merge an empty SIR EU list");
-    if units.len() == 1 {
-        return (
-            (*units[0]).clone(),
-            SirMergeProvenance {
-                unit_entries: vec![units[0].entry_block_id],
-                block_units: units[0]
-                    .blocks
-                    .keys()
-                    .copied()
-                    .map(|block| (block, 0))
-                    .collect(),
-            },
-        );
-    }
-
-    let mut merged_blocks = crate::HashMap::default();
-    let mut merged_regs = crate::HashMap::default();
-    let mut block_units = crate::HashMap::default();
-
-    // Compute offsets for renumbering
-    let mut reg_offset = 0usize;
-    let mut block_offset = 0usize;
-    let mut reg_offsets = Vec::new();
-    let mut block_offsets = Vec::new();
-
-    for eu in units {
-        reg_offsets.push(reg_offset);
-        block_offsets.push(block_offset);
-        let max_reg = eu.register_map.keys().map(|r| r.0).max().unwrap_or(0);
-        reg_offset += max_reg + 1;
-        let max_block = eu.blocks.keys().map(|b| b.0).max().unwrap_or(0);
-        block_offset += max_block + 1;
-    }
-
-    // Entry block IDs after renumbering
-    let entry_blocks: Vec<BlockId> = units
-        .iter()
-        .enumerate()
-        .map(|(i, eu)| BlockId(eu.entry_block_id.0 + block_offsets[i]))
-        .collect();
-
-    for (eu_idx, eu) in units.iter().enumerate() {
-        let ro = reg_offsets[eu_idx];
-        let bo = block_offsets[eu_idx];
-        let is_last = eu_idx == units.len() - 1;
-        let next_entry = if !is_last {
-            Some(entry_blocks[eu_idx + 1])
-        } else {
-            None
-        };
-
-        // Copy registers with offset
-        for (&reg_id, reg_type) in &eu.register_map {
-            merged_regs.insert(RegisterId(reg_id.0 + ro), reg_type.clone());
-        }
-
-        // Copy blocks with renumbering
-        for (&block_id, block) in &eu.blocks {
-            let new_block_id = BlockId(block_id.0 + bo);
-            block_units.insert(new_block_id, eu_idx);
-            let r = |reg: RegisterId| RegisterId(reg.0 + ro);
-            let b = |bid: BlockId| BlockId(bid.0 + bo);
-
-            let new_params: Vec<RegisterId> = block.params.iter().map(|p| r(*p)).collect();
-            let new_insts: Vec<SIRInstruction<A>> = block
-                .instructions
-                .iter()
-                .map(|inst| renumber_sir_inst(inst, ro, bo))
-                .collect();
-
-            let new_terminator = match &block.terminator {
-                SIRTerminator::Return => {
-                    if is_last {
-                        SIRTerminator::Return
-                    } else {
-                        SIRTerminator::Jump(next_entry.unwrap(), vec![])
-                    }
-                }
-                SIRTerminator::Error(code) => SIRTerminator::Error(*code),
-                SIRTerminator::Jump(target, args) => {
-                    let new_args: Vec<RegisterId> = args.iter().map(|a| r(*a)).collect();
-                    SIRTerminator::Jump(b(*target), new_args)
-                }
-                SIRTerminator::Branch {
-                    cond,
-                    true_block,
-                    false_block,
-                } => SIRTerminator::Branch {
-                    cond: r(*cond),
-                    true_block: (
-                        b(true_block.0),
-                        true_block.1.iter().map(|a| r(*a)).collect(),
-                    ),
-                    false_block: (
-                        b(false_block.0),
-                        false_block.1.iter().map(|a| r(*a)).collect(),
-                    ),
-                },
-                SIRTerminator::Switch {
-                    selector,
-                    cases,
-                    default,
-                } => SIRTerminator::Switch {
-                    selector: r(*selector),
-                    cases: cases
-                        .iter()
-                        .map(|case| SIRSwitchCase {
-                            value: case.value.clone(),
-                            target: b(case.target),
-                        })
-                        .collect(),
-                    default: b(*default),
-                },
-            };
-
-            merged_blocks.insert(
-                new_block_id,
-                BasicBlock {
-                    id: new_block_id,
-                    params: new_params,
-                    instructions: new_insts,
-                    terminator: new_terminator,
-                },
-            );
-        }
-    }
-
-    (
-        ExecutionUnit {
-            entry_block_id: entry_blocks[0],
-            blocks: merged_blocks,
-            register_map: merged_regs,
-        },
-        SirMergeProvenance {
-            unit_entries: entry_blocks,
-            block_units,
-        },
-    )
-}
-
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-pub(crate) fn inline_single_predecessor_jumps<A: Clone>(
-    eu: &mut ExecutionUnit<A>,
-) -> Result<bool, crate::ir::verify::SirVerifyError> {
-    fn successors(terminator: &SIRTerminator) -> Vec<BlockId> {
-        match terminator {
-            SIRTerminator::Jump(target, _) => vec![*target],
-            SIRTerminator::Branch {
-                true_block,
-                false_block,
-                ..
-            } => vec![true_block.0, false_block.0],
-            SIRTerminator::Switch { cases, default, .. } => cases
-                .iter()
-                .map(|case| case.target)
-                .chain(std::iter::once(*default))
-                .collect(),
-            SIRTerminator::Return | SIRTerminator::Error(_) => Vec::new(),
-        }
-    }
-
-    let mut changed = false;
-    let mut parameter_replacements = crate::HashMap::<RegisterId, RegisterId>::default();
-    let mut predecessor_count = crate::HashMap::<BlockId, usize>::default();
-    for block in eu.blocks.values() {
-        for successor in successors(&block.terminator) {
-            *predecessor_count.entry(successor).or_default() += 1;
-        }
-    }
-
-    // Inlining `source -> target` replaces every outgoing edge of `target`
-    // with the corresponding edge from `source`. Therefore predecessor edge
-    // counts of all surviving blocks stay unchanged. Only `source` gets a new
-    // terminator and can become a new candidate. A deterministic worklist
-    // avoids rebuilding and sorting the whole CFG after every inlining step.
-    let mut candidates = eu.blocks.keys().copied().collect::<BTreeSet<_>>();
-    while let Some(block_id) = candidates.pop_first() {
-        let Some(block) = eu.blocks.get(&block_id) else {
-            continue;
-        };
-        let SIRTerminator::Jump(target, args) = &block.terminator else {
-            continue;
-        };
-        if *target == block_id
-            || *target == eu.entry_block_id
-            || predecessor_count.get(target).copied().unwrap_or(0) != 1
-        {
-            continue;
-        }
-        let Some(target_block) = eu.blocks.get(target) else {
-            continue;
-        };
-        if target_block.params.len() != args.len() {
-            continue;
-        }
-        let target_id = *target;
-        let args = args.clone();
-
-        let target = eu
-            .blocks
-            .remove(&target_id)
-            .expect("inline target exists when selected");
-        candidates.remove(&target_id);
-        for (parameter, argument) in target.params.iter().copied().zip(args) {
-            parameter_replacements.insert(parameter, argument);
-        }
-
-        let block = eu
-            .blocks
-            .get_mut(&block_id)
-            .expect("inline predecessor exists when selected");
-        block.instructions.extend(target.instructions);
-        block.terminator = target.terminator;
-        candidates.insert(block_id);
-        changed = true;
-    }
-
-    // A block parameter is an SSA definition whose uses may appear in any
-    // block dominated by the removed block. Resolve chains of removed
-    // parameters once, then rewrite the surviving unit in one linear pass.
-    let mut flattened = crate::HashMap::<RegisterId, RegisterId>::default();
-    let mut parameters = parameter_replacements.keys().copied().collect::<Vec<_>>();
-    parameters.sort_unstable();
-    for parameter in parameters {
-        if flattened.contains_key(&parameter) {
-            continue;
-        }
-        let mut path = Vec::new();
-        let mut on_path = crate::HashSet::default();
-        let mut current = parameter;
-        let resolved = loop {
-            if let Some(&resolved) = flattened.get(&current) {
-                break resolved;
-            }
-            let Some(&next) = parameter_replacements.get(&current) else {
-                break current;
-            };
-            if !on_path.insert(current) {
-                return Err(crate::ir::verify::SirVerifyError {
-                    invariant: "SSA.INLINE_PARAMETER_ACYCLIC",
-                    block: None,
-                    instruction: None,
-                    message: format!("single-predecessor parameter cycle reaches r{}", current.0),
-                });
-            }
-            path.push(current);
-            current = next;
-        };
-        for register in path {
-            flattened.insert(register, resolved);
-        }
-    }
-    if !flattened.is_empty() {
-        for block in eu.blocks.values_mut() {
-            for instruction in &mut block.instructions {
-                replace_sir_uses(instruction, &flattened);
-            }
-            replace_sir_terminator_uses(&mut block.terminator, &flattened);
-        }
-    }
-    if let Err(mut error) = eu.verify_result() {
-        error.message = format!("after single-predecessor inlining: {}", error.message);
-        return Err(error);
-    }
-    Ok(changed)
-}
-
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-fn replace_sir_offset_uses(
-    offset: &mut SIROffset,
-    replacements: &crate::HashMap<RegisterId, RegisterId>,
-) {
-    match offset {
-        SIROffset::Static(_) | SIROffset::PackedElements { .. } => {}
-        SIROffset::Dynamic(register) => {
-            if let Some(&replacement) = replacements.get(register) {
-                *register = replacement;
-            }
-        }
-        SIROffset::Element {
-            index,
-            dynamic_bit_offset,
-            ..
-        } => {
-            if let Some(&replacement) = replacements.get(index) {
-                *index = replacement;
-            }
-            if let Some(register) = dynamic_bit_offset {
-                if let Some(&replacement) = replacements.get(register) {
-                    *register = replacement;
-                }
-            }
-        }
-    }
-}
-
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-fn replace_sir_uses<A>(
-    instruction: &mut SIRInstruction<A>,
-    replacements: &crate::HashMap<RegisterId, RegisterId>,
-) {
-    let replace = |register: &mut RegisterId| {
-        if let Some(&replacement) = replacements.get(register) {
-            *register = replacement;
-        }
-    };
-    match instruction {
-        SIRInstruction::Imm(..) => {}
-        SIRInstruction::Binary(_, lhs, _, rhs) => {
-            replace(lhs);
-            replace(rhs);
-        }
-        SIRInstruction::Unary(_, _, source) | SIRInstruction::Slice(_, source, _, _) => {
-            replace(source);
-        }
-        SIRInstruction::Load(_, _, offset, _) => {
-            replace_sir_offset_uses(offset, replacements);
-        }
-        SIRInstruction::Store(_, offset, _, source, _, _) => {
-            replace_sir_offset_uses(offset, replacements);
-            replace(source);
-        }
-        SIRInstruction::Commit(_, _, offset, _, _) => {
-            replace_sir_offset_uses(offset, replacements);
-        }
-        SIRInstruction::Concat(_, sources) => {
-            for source in sources {
-                replace(source);
-            }
-        }
-        SIRInstruction::Mux(_, condition, then_value, else_value) => {
-            replace(condition);
-            replace(then_value);
-            replace(else_value);
-        }
-        SIRInstruction::RuntimeEvent { args, .. } => {
-            for arg in args {
-                replace(arg);
-            }
-        }
-        SIRInstruction::CombCaptureEvent { args, .. } => {
-            for arg in args {
-                replace(arg);
-            }
-        }
-        SIRInstruction::CombCaptureEnableIfChanged { old, new, .. } => {
-            replace(old);
-            replace(new);
-        }
-    }
-}
-
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-fn replace_sir_terminator_uses(
-    terminator: &mut SIRTerminator,
-    replacements: &crate::HashMap<RegisterId, RegisterId>,
-) {
-    let replace = |register: &mut RegisterId| {
-        if let Some(&replacement) = replacements.get(register) {
-            *register = replacement;
-        }
-    };
-    match terminator {
-        SIRTerminator::Jump(_, args) => {
-            for arg in args {
-                replace(arg);
-            }
-        }
-        SIRTerminator::Branch {
-            cond,
-            true_block,
-            false_block,
-        } => {
-            replace(cond);
-            for arg in &mut true_block.1 {
-                replace(arg);
-            }
-            for arg in &mut false_block.1 {
-                replace(arg);
-            }
-        }
-        SIRTerminator::Switch {
-            selector,
-            cases: _,
-            default: _,
-        } => {
-            replace(selector);
-        }
-        SIRTerminator::Return | SIRTerminator::Error(_) => {}
-    }
-}
-
-#[cfg_attr(not(target_arch = "x86_64"), allow(dead_code))]
-fn renumber_sir_inst<A: Clone>(
-    inst: &SIRInstruction<A>,
-    ro: usize,
-    _bo: usize,
-) -> SIRInstruction<A> {
-    let r = |reg: RegisterId| RegisterId(reg.0 + ro);
-    let off = |o: &SIROffset| match o {
-        SIROffset::Static(v) => SIROffset::Static(*v),
-        SIROffset::Dynamic(reg) => SIROffset::Dynamic(r(*reg)),
-        SIROffset::Element {
-            index,
-            element_width,
-            bit_offset,
-            dynamic_bit_offset,
-        } => SIROffset::Element {
-            index: r(*index),
-            element_width: *element_width,
-            bit_offset: *bit_offset,
-            dynamic_bit_offset: dynamic_bit_offset.map(r),
-        },
-        SIROffset::PackedElements {
-            bit_offset,
-            element_width,
-        } => SIROffset::PackedElements {
-            bit_offset: *bit_offset,
-            element_width: *element_width,
-        },
-    };
-
-    match inst {
-        SIRInstruction::Imm(dst, val) => SIRInstruction::Imm(r(*dst), val.clone()),
-        SIRInstruction::Load(dst, addr, offset, width) => {
-            SIRInstruction::Load(r(*dst), addr.clone(), off(offset), *width)
-        }
-        SIRInstruction::Store(addr, offset, width, src, triggers, comb_capture_sites) => {
-            SIRInstruction::Store(
-                addr.clone(),
-                off(offset),
-                *width,
-                r(*src),
-                triggers.clone(),
-                comb_capture_sites.clone(),
-            )
-        }
-        SIRInstruction::Commit(src, dst, offset, width, triggers) => SIRInstruction::Commit(
-            src.clone(),
-            dst.clone(),
-            off(offset),
-            *width,
-            triggers.clone(),
-        ),
-        SIRInstruction::Binary(dst, lhs, op, rhs) => {
-            SIRInstruction::Binary(r(*dst), r(*lhs), *op, r(*rhs))
-        }
-        SIRInstruction::Unary(dst, op, src) => SIRInstruction::Unary(r(*dst), *op, r(*src)),
-        SIRInstruction::Concat(dst, args) => {
-            SIRInstruction::Concat(r(*dst), args.iter().map(|a| r(*a)).collect())
-        }
-        SIRInstruction::Slice(dst, src, offset, width) => {
-            SIRInstruction::Slice(r(*dst), r(*src), *offset, *width)
-        }
-        SIRInstruction::Mux(dst, cond, then_val, else_val) => {
-            SIRInstruction::Mux(r(*dst), r(*cond), r(*then_val), r(*else_val))
-        }
-        SIRInstruction::RuntimeEvent { site_id, args } => SIRInstruction::RuntimeEvent {
-            site_id: *site_id,
-            args: args.iter().map(|a| r(*a)).collect(),
-        },
-        SIRInstruction::CombCaptureEvent {
-            site_id,
-            args,
-            fatal_error_code,
-            consume_enabled,
-        } => SIRInstruction::CombCaptureEvent {
-            site_id: *site_id,
-            args: args.iter().map(|a| r(*a)).collect(),
-            fatal_error_code: *fatal_error_code,
-            consume_enabled: *consume_enabled,
-        },
-        SIRInstruction::CombCaptureEnableIfChanged { old, new, sites } => {
-            SIRInstruction::CombCaptureEnableIfChanged {
-                old: r(*old),
-                new: r(*new),
-                sites: sites.clone(),
-            }
-        }
-    }
-}
-
-/// Basic Block: A sequence of linear instructions and a terminator instruction
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(bound(serialize = "Addr: Serialize", deserialize = "Addr: Deserialize<'de>"))]
-pub struct BasicBlock<Addr> {
-    pub id: BlockId,
-    pub params: Vec<RegisterId>,
-    /// List of side-effect-free operations, Loads, and Stores
-    pub instructions: Vec<SIRInstruction<Addr>>,
-    /// Where to transition at the end of this block (key for short-circuit evaluation)
-    pub terminator: SIRTerminator,
-}
-
-impl<A: Display> fmt::Display for BasicBlock<A> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        writeln!(f, "b{}:", self.id.0)?;
-        if !self.params.is_empty() {
-            write!(f, "  params: [")?;
-            for (i, param) in self.params.iter().enumerate() {
-                if i > 0 {
-                    write!(f, ", ")?;
-                }
-                write!(f, "r{}", param.0)?;
-            }
-            writeln!(f, "]")?;
-        }
-        for inst in &self.instructions {
-            writeln!(f, "  {}", inst)?;
-        }
-        write!(f, "  {}", self.terminator)
-    }
-}
-
-/// Terminator instruction: Determines control flow
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum SIRTerminator {
-    /// Unconditional transition to the next block
-    Jump(BlockId, Vec<RegisterId>),
-    /// Conditional branch (true_block if cond is non-zero, false_block if zero)
-    Branch {
-        cond: RegisterId,
-        true_block: (BlockId, Vec<RegisterId>),
-        false_block: (BlockId, Vec<RegisterId>),
-    },
-    /// Exact multiway dispatch for a selector of at most eight bits. Cases are
-    /// tested against the declared bit width; values not listed take `default`.
-    Switch {
-        selector: RegisterId,
-        cases: Vec<SIRSwitchCase>,
-        default: BlockId,
-    },
-    /// End of module execution
-    Return,
-    Error(i64),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct SIRSwitchCase {
-    #[serde(with = "crate::serde_helpers::biguint")]
-    pub value: BigUint,
-    pub target: BlockId,
-}
-
-impl fmt::Display for SIRTerminator {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // Helper to format block ID and argument list
-        let fmt_target =
-            |f: &mut fmt::Formatter<'_>, id: BlockId, args: &[RegisterId]| -> fmt::Result {
-                write!(f, "b{}", id.0)?;
-                if !args.is_empty() {
-                    write!(f, " [")?;
-                    for (i, arg) in args.iter().enumerate() {
-                        if i > 0 {
-                            write!(f, ", ")?;
-                        }
-                        write!(f, "r{}", arg.0)?;
-                    }
-                    write!(f, "]")?;
-                }
-                Ok(())
-            };
-
-        match self {
-            SIRTerminator::Jump(block_id, args) => {
-                write!(f, "Jump(")?;
-                fmt_target(f, *block_id, args)?;
-                write!(f, ")")
-            }
-            SIRTerminator::Branch {
-                cond,
-                true_block,
-                false_block,
-            } => {
-                write!(f, "Branch(r{} ? ", cond.0)?;
-                fmt_target(f, true_block.0, &true_block.1)?;
-                write!(f, " : ")?;
-                fmt_target(f, false_block.0, &false_block.1)?;
-                write!(f, ")")
-            }
-            SIRTerminator::Switch {
-                selector,
-                cases,
-                default,
-            } => {
-                write!(f, "Switch(r{}; ", selector.0)?;
-                for (index, case) in cases.iter().enumerate() {
-                    if index != 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{:#x} => ", case.value)?;
-                    fmt_target(f, case.target, &[])?;
-                }
-                write!(f, "; default => ")?;
-                fmt_target(f, *default, &[])?;
-                write!(f, ")")
-            }
-            SIRTerminator::Return => write!(f, "Return"),
-            SIRTerminator::Error(code) => write!(f, "Error({})", code),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum RegisterType {
-    Logic { width: usize },
-    Bit { width: usize, signed: bool },
-}
-impl RegisterType {
-    pub fn width(&self) -> usize {
-        match self {
-            RegisterType::Bit { width, signed: _ } => *width,
-            RegisterType::Logic { width } => *width,
-        }
-    }
-    pub fn is_signed(&self) -> bool {
-        matches!(
-            self,
-            RegisterType::Bit {
-                width: _,
-                signed: true
-            }
-        )
-    }
-}
-
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub struct RegisterId(pub usize);
-
-impl fmt::Display for RegisterId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "r{}", self.0)
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum BinaryOp {
-    Add,
-    Sub,
-    Mul,
-    DivU,
-    DivS,
-    RemU,
-    RemS,
-    And,
-    Or,
-    Xor,
-    Shl, // Logical Shift Left (<<)
-    Shr, // Logical Shift Right (>>)
-    Sar, // Arithmetic Shift Right (>>>)
-    Eq,
-    Ne,
-    LtU,
-    LtS, // Less Than (Unsigned / Signed)
-    LeU,
-    LeS, // Less Equal
-    GtU,
-    GtS, // Greater Than
-    GeU,
-    GeS, // Greater Equal
-    LogicAnd,
-    LogicOr,
-    EqWildcard,
-    NeWildcard,
-}
-
-impl BinaryOp {
-    /// Whether the operation is commutative (a op b == b op a).
-    pub fn is_commutative(&self) -> bool {
-        matches!(
-            self,
-            BinaryOp::Add
-                | BinaryOp::Mul
-                | BinaryOp::And
-                | BinaryOp::Or
-                | BinaryOp::Xor
-                | BinaryOp::Eq
-                | BinaryOp::Ne
-                | BinaryOp::LogicAnd
-                | BinaryOp::LogicOr
-        )
-    }
-}
-
-impl fmt::Display for BinaryOp {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let op_str = match self {
-            BinaryOp::Add => "Add",
-            BinaryOp::Sub => "Sub",
-            BinaryOp::Mul => "Mul",
-            BinaryOp::DivU => "DivU",
-            BinaryOp::DivS => "DivS",
-            BinaryOp::RemU => "RemU",
-            BinaryOp::RemS => "RemS",
-            BinaryOp::And => "And",
-            BinaryOp::Or => "Or",
-            BinaryOp::Xor => "Xor",
-            BinaryOp::Shl => "Shl",
-            BinaryOp::Shr => "Shr",
-            BinaryOp::Sar => "Sar",
-            BinaryOp::Eq => "Eq",
-            BinaryOp::Ne => "Ne",
-            BinaryOp::LtU => "LtU",
-            BinaryOp::LtS => "LtS",
-            BinaryOp::LeU => "LeU",
-            BinaryOp::LeS => "LeS",
-            BinaryOp::GtU => "GtU",
-            BinaryOp::GtS => "GtS",
-            BinaryOp::GeU => "GeU",
-            BinaryOp::GeS => "GeS",
-            BinaryOp::LogicAnd => "LogicAnd",
-            BinaryOp::LogicOr => "LogicOr",
-            BinaryOp::EqWildcard => "EqWildcard",
-            BinaryOp::NeWildcard => "NeWildcard",
-        };
-        write!(f, "{}", op_str)
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum UnaryOp {
-    Ident,
-    /// Convert a four-state value to two-state form. Unknown bits become zero:
-    /// `(value, mask) -> (value & !mask, 0)`.
-    ToTwoState,
-    Minus,
-    BitNot,
-    LogicNot,
-    And,
-    Or,
-    Xor,
-    PopCount,
-    CountLeadingZeros,
-    CountTrailingZeros,
-}
-
-impl UnaryOp {
-    /// Return the canonical result width for an operand of `operand_width` bits.
-    ///
-    /// Bit-count operations return a value in `0..=operand_width`, which needs
-    /// `ceil(log2(operand_width + 1))` bits.  Computing that as the bit length
-    /// of `operand_width` avoids overflowing when the operand width is
-    /// `usize::MAX`.
-    pub fn result_width(self, operand_width: usize) -> usize {
-        match self {
-            UnaryOp::LogicNot | UnaryOp::And | UnaryOp::Or | UnaryOp::Xor => 1,
-            UnaryOp::Ident | UnaryOp::ToTwoState | UnaryOp::Minus | UnaryOp::BitNot => {
-                operand_width
-            }
-            UnaryOp::PopCount | UnaryOp::CountLeadingZeros | UnaryOp::CountTrailingZeros => {
-                usize::BITS as usize - operand_width.leading_zeros() as usize
-            }
-        }
-    }
-}
-
-impl fmt::Display for UnaryOp {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let op_str = match self {
-            UnaryOp::Ident => "Ident",
-            UnaryOp::ToTwoState => "ToTwoState",
-            UnaryOp::Minus => "Minus",
-            UnaryOp::BitNot => "BitNot",
-            UnaryOp::LogicNot => "LogicNot",
-            UnaryOp::And => "And",
-            UnaryOp::Or => "Or",
-            UnaryOp::Xor => "Xor",
-            UnaryOp::PopCount => "PopCount",
-            UnaryOp::CountLeadingZeros => "CountLeadingZeros",
-            UnaryOp::CountTrailingZeros => "CountTrailingZeros",
-        };
-        write!(f, "{}", op_str)
-    }
-}
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-
-pub struct SIRValue {
-    #[serde(with = "crate::serde_helpers::biguint")]
-    pub payload: BigUint,
-    #[serde(with = "crate::serde_helpers::biguint")]
-    pub mask: BigUint,
-}
-impl SIRValue {
-    pub fn new(payload: impl Into<BigUint>) -> Self {
-        Self {
-            payload: payload.into(),
-            mask: BigUint::from(0u32),
-        }
-    }
-    pub fn new_four_state(payload: impl Into<BigUint>, mask: impl Into<BigUint>) -> Self {
-        Self {
-            payload: payload.into(),
-            mask: mask.into(),
-        }
-    }
-}
-
-impl fmt::Display for SIRValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if self.mask == BigUint::from(0u32) {
-            write!(f, "SIRValue({:#x})", self.payload)
-        } else {
-            write!(f, "SIRValue({:#x}, mask={:#x})", self.payload, self.mask)
-        }
-    }
-}
-
-#[derive(Debug, Clone, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-pub enum SIROffset {
-    /// Static bit offset
-    Static(usize),
-    /// Dynamic bit offset (register value)
-    Dynamic(RegisterId),
-    /// Access to one element of an unpacked array.
-    ///
-    /// `index` is the flattened element index, not a bit offset.  The logical
-    /// bit offset is `index * element_width + bit_offset`.  Keeping this form
-    /// in SIR lets a backend choose an element-strided physical layout without
-    /// recovering source type information from arithmetic instructions.
-    Element {
-        index: RegisterId,
-        element_width: usize,
-        bit_offset: usize,
-        dynamic_bit_offset: Option<RegisterId>,
-    },
-    /// A vectorized read or write of consecutive unpacked elements through
-    /// their packed logical representation.
-    ///
-    /// Unlike `Static`, this explicitly permits crossing element boundaries.
-    /// A backend may require the addressed object to use packed storage.
-    PackedElements {
-        bit_offset: usize,
-        element_width: usize,
-    },
-}
-
-impl fmt::Display for SIROffset {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            SIROffset::Static(val) => write!(f, "{}", val),
-            SIROffset::Dynamic(reg) => write!(f, "r{}", reg.0),
-            SIROffset::Element {
-                index,
-                element_width,
-                bit_offset,
-                dynamic_bit_offset,
-            } => {
-                write!(
-                    f,
-                    "element(r{}, width={}, bit={}",
-                    index.0, element_width, bit_offset
-                )?;
-                if let Some(dynamic) = dynamic_bit_offset {
-                    write!(f, "+r{}", dynamic.0)?;
-                }
-                write!(f, ")")
-            }
-            SIROffset::PackedElements {
-                bit_offset,
-                element_width,
-            } => write!(
-                f,
-                "packed_elements(bit={}, element_width={})",
-                bit_offset, element_width
-            ),
-        }
-    }
-}
-
-impl SIROffset {
-    /// Returns the constant offset in the object's packed logical bit space.
-    ///
-    /// `PackedElements` differs from `Static` in the physical layouts a
-    /// backend may select, but both name an exact logical range. Analyses
-    /// which only reason about aliasing must not treat `PackedElements` as a
-    /// dynamic or unknown access.
-    pub fn constant_bit_offset(&self) -> Option<usize> {
-        match self {
-            SIROffset::Static(bit_offset) | SIROffset::PackedElements { bit_offset, .. } => {
-                Some(*bit_offset)
-            }
-            SIROffset::Dynamic(_) | SIROffset::Element { .. } => None,
-        }
-    }
-
-    pub fn dynamic_registers(&self) -> [Option<RegisterId>; 2] {
-        match self {
-            SIROffset::Static(_) => [None, None],
-            SIROffset::Dynamic(register) => [Some(*register), None],
-            SIROffset::Element {
-                index,
-                dynamic_bit_offset,
-                ..
-            } => [Some(*index), *dynamic_bit_offset],
-            SIROffset::PackedElements { .. } => [None, None],
-        }
-    }
-
-    pub fn is_dynamic(&self) -> bool {
-        matches!(self, SIROffset::Dynamic(_) | SIROffset::Element { .. })
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(bound(serialize = "Addr: Serialize", deserialize = "Addr: Deserialize<'de>"))]
-pub enum SIRInstruction<Addr> {
-    Imm(RegisterId, SIRValue),
-    Binary(RegisterId, RegisterId, BinaryOp, RegisterId),
-    Unary(RegisterId, UnaryOp, RegisterId),
-    Load(RegisterId, Addr, SIROffset, usize),
-    Store(
-        Addr,
-        SIROffset,
-        usize,
-        RegisterId,
-        Vec<TriggerIdWithKind>,
-        Vec<u32>,
-    ),
-    /// Commits a value from `src` region to `dst` region with the same offset/width.
-    Commit(Addr, Addr, SIROffset, usize, Vec<TriggerIdWithKind>),
-    /// Concatenates multiple registers into a single register.
-    /// Order: [MSB, ..., LSB] (First element is most significant)
-    Concat(RegisterId, Vec<RegisterId>),
-    /// Extracts a bit range from a register: dst = src[offset +: width].
-    /// Lowered to O(1) chunk-indexed loads in the CLIF backend.
-    Slice(RegisterId, RegisterId, usize, usize), // dst, src, bit_offset, width
-    /// Mux: dst = if cond { then_val } else { else_val }.
-    /// In 4-state mode, preserves exact mask bits (including Z) of the selected branch.
-    /// A known one in `cond` selects `then_val`. If `cond` has no known one but
-    /// contains X/Z, equal arm bits are preserved and differing bits become X.
-    Mux(RegisterId, RegisterId, RegisterId, RegisterId), // dst, cond, then_val, else_val
-    RuntimeEvent {
-        site_id: u32,
-        args: Vec<RegisterId>,
-    },
-    CombCaptureEvent {
-        site_id: u32,
-        args: Vec<RegisterId>,
-        fatal_error_code: Option<i64>,
-        consume_enabled: bool,
-    },
-    CombCaptureEnableIfChanged {
-        old: RegisterId,
-        new: RegisterId,
-        sites: Vec<u32>,
-    },
-}
-
-impl<A: Display> fmt::Display for SIRInstruction<A> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            SIRInstruction::Imm(rd, value) => {
-                write!(f, "r{} = {}", rd.0, value)
-            }
-            SIRInstruction::Binary(rd, rs1, op, rs2) => {
-                write!(f, "r{} = r{} {} r{}", rd.0, rs1.0, op, rs2.0)
-            }
-            SIRInstruction::Unary(rd, op, rs) => {
-                write!(f, "r{} = {} r{}", rd.0, op, rs.0)
-            }
-            SIRInstruction::Load(rd, addr, offset, bits) => {
-                write!(
-                    f,
-                    "r{} = Load(addr={}, offset={}, bits={})",
-                    rd.0, addr, offset, bits
-                )
-            }
-            SIRInstruction::Store(
-                addr,
-                offset,
-                op_width,
-                src_reg,
-                triggers,
-                comb_capture_sites,
-            ) => {
-                write!(
-                    f,
-                    "Store(addr={}, offset={}, src_reg = {}, bits={}, triggers={:?}, comb_capture_sites={:?})",
-                    addr, offset, src_reg.0, op_width, triggers, comb_capture_sites
-                )
-            }
-            SIRInstruction::Commit(src, dst, offset, bits, triggers) => {
-                write!(
-                    f,
-                    "Commit(src={}, dst={}, offset={}, bits={}, triggers={:?})",
-                    src, dst, offset, bits, triggers
-                )
-            }
-            SIRInstruction::Concat(dst, args) => {
-                write!(f, "r{} = Concat([", dst.0)?;
-                for (i, arg) in args.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "r{}", arg.0)?;
-                }
-                write!(f, "])")
-            }
-            SIRInstruction::Slice(dst, src, offset, width) => {
-                write!(
-                    f,
-                    "r{} = Slice(r{}, offset={}, width={})",
-                    dst.0, src.0, offset, width
-                )
-            }
-            SIRInstruction::Mux(dst, cond, then_val, else_val) => {
-                write!(
-                    f,
-                    "r{} = Mux(cond=r{}, then=r{}, else=r{})",
-                    dst.0, cond.0, then_val.0, else_val.0
-                )
-            }
-            SIRInstruction::RuntimeEvent { site_id, args } => {
-                write!(f, "RuntimeEvent(site={}, args=[", site_id)?;
-                for (i, arg) in args.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "r{}", arg.0)?;
-                }
-                write!(f, "])")
-            }
-            SIRInstruction::CombCaptureEvent {
-                site_id,
-                args,
-                fatal_error_code,
-                consume_enabled,
-            } => {
-                write!(f, "CombCaptureEvent(site={}, args=[", site_id)?;
-                for (i, arg) in args.iter().enumerate() {
-                    if i > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "r{}", arg.0)?;
-                }
-                if let Some(code) = fatal_error_code {
-                    write!(f, "], fatal_error={code})")
-                } else if *consume_enabled {
-                    write!(f, "], consume_enabled=true)")
-                } else {
-                    write!(f, "])")
-                }
-            }
-            SIRInstruction::CombCaptureEnableIfChanged { old, new, sites } => {
-                write!(
-                    f,
-                    "CombCaptureEnableIfChanged(old=r{}, new=r{}, sites={:?})",
-                    old.0, new.0, sites
-                )
-            }
-        }
-    }
-}
-impl<A> SIRInstruction<A> {
-    pub(crate) fn defined_register(&self) -> Option<RegisterId> {
-        match self {
-            SIRInstruction::Imm(dst, _)
-            | SIRInstruction::Binary(dst, _, _, _)
-            | SIRInstruction::Unary(dst, _, _)
-            | SIRInstruction::Load(dst, _, _, _)
-            | SIRInstruction::Concat(dst, _)
-            | SIRInstruction::Slice(dst, _, _, _)
-            | SIRInstruction::Mux(dst, _, _, _) => Some(*dst),
-            SIRInstruction::Store(..)
-            | SIRInstruction::Commit(..)
-            | SIRInstruction::RuntimeEvent { .. }
-            | SIRInstruction::CombCaptureEvent { .. }
-            | SIRInstruction::CombCaptureEnableIfChanged { .. } => None,
-        }
-    }
-
-    pub fn into_map_addr<B>(self, mut f: impl FnMut(A) -> B) -> SIRInstruction<B> {
-        match self {
-            SIRInstruction::Imm(register_id, value) => SIRInstruction::Imm(register_id, value),
-            SIRInstruction::Binary(rd, rs1, op, rs2) => SIRInstruction::Binary(rd, rs1, op, rs2),
-            SIRInstruction::Unary(rd, op, rs) => SIRInstruction::Unary(rd, op, rs),
-            SIRInstruction::Load(rd, addr, offset, bits) => {
-                SIRInstruction::Load(rd, f(addr), offset, bits)
-            }
-            SIRInstruction::Store(addr, offset, bits, rs, triggers, comb_capture_sites) => {
-                SIRInstruction::Store(f(addr), offset, bits, rs, triggers, comb_capture_sites)
-            }
-            SIRInstruction::Commit(src, dst, offset, bits, triggers) => {
-                SIRInstruction::Commit(f(src), f(dst), offset, bits, triggers)
-            }
-            SIRInstruction::Concat(dst, args) => SIRInstruction::Concat(dst, args),
-            SIRInstruction::Slice(dst, src, offset, width) => {
-                SIRInstruction::Slice(dst, src, offset, width)
-            }
-            SIRInstruction::Mux(dst, cond, then_val, else_val) => {
-                SIRInstruction::Mux(dst, cond, then_val, else_val)
-            }
-            SIRInstruction::RuntimeEvent { site_id, args } => {
-                SIRInstruction::RuntimeEvent { site_id, args }
-            }
-            SIRInstruction::CombCaptureEvent {
-                site_id,
-                args,
-                fatal_error_code,
-                consume_enabled,
-            } => SIRInstruction::CombCaptureEvent {
-                site_id,
-                args,
-                fatal_error_code,
-                consume_enabled,
-            },
-            SIRInstruction::CombCaptureEnableIfChanged { old, new, sites } => {
-                SIRInstruction::CombCaptureEnableIfChanged { old, new, sites }
-            }
-        }
-    }
-    pub fn map_addr<B>(&self, mut f: impl FnMut(&A) -> B) -> SIRInstruction<B> {
-        match self {
-            SIRInstruction::Imm(register_id, value) => {
-                SIRInstruction::Imm(*register_id, value.clone())
-            }
-            SIRInstruction::Binary(rd, rs1, op, rs2) => {
-                SIRInstruction::Binary(*rd, *rs1, *op, *rs2)
-            }
-            SIRInstruction::Unary(rd, op, rs) => SIRInstruction::Unary(*rd, *op, *rs),
-            SIRInstruction::Load(rd, addr, offset, bits) => {
-                SIRInstruction::Load(*rd, f(addr), offset.clone(), *bits)
-            }
-            SIRInstruction::Store(addr, offset, bits, rs, triggers, comb_capture_sites) => {
-                SIRInstruction::Store(
-                    f(addr),
-                    offset.clone(),
-                    *bits,
-                    *rs,
-                    triggers.clone(),
-                    comb_capture_sites.clone(),
-                )
-            }
-            SIRInstruction::Commit(src, dst, offset, bits, triggers) => {
-                SIRInstruction::Commit(f(src), f(dst), offset.clone(), *bits, triggers.clone())
-            }
-            SIRInstruction::Concat(dst, args) => SIRInstruction::Concat(*dst, args.clone()),
-            SIRInstruction::Slice(dst, src, offset, width) => {
-                SIRInstruction::Slice(*dst, *src, *offset, *width)
-            }
-            SIRInstruction::Mux(dst, cond, then_val, else_val) => {
-                SIRInstruction::Mux(*dst, *cond, *then_val, *else_val)
-            }
-            SIRInstruction::RuntimeEvent { site_id, args } => SIRInstruction::RuntimeEvent {
-                site_id: *site_id,
-                args: args.clone(),
-            },
-            SIRInstruction::CombCaptureEvent {
-                site_id,
-                args,
-                fatal_error_code,
-                consume_enabled,
-            } => SIRInstruction::CombCaptureEvent {
-                site_id: *site_id,
-                args: args.clone(),
-                fatal_error_code: *fatal_error_code,
-                consume_enabled: *consume_enabled,
-            },
-            SIRInstruction::CombCaptureEnableIfChanged { old, new, sites } => {
-                SIRInstruction::CombCaptureEnableIfChanged {
-                    old: *old,
-                    new: *new,
-                    sites: sites.clone(),
-                }
-            }
-        }
-    }
-}
-
-fn visit_exact_zero_dependencies<A>(
-    instruction: &SIRInstruction<A>,
-    mut visit: impl FnMut(RegisterId),
-) -> Option<usize> {
-    let mut count = 0usize;
-    let mut dependency = |register| {
-        count += 1;
-        visit(register);
-    };
-    match instruction {
-        SIRInstruction::Binary(
-            _,
-            lhs,
-            BinaryOp::Add
-            | BinaryOp::Sub
-            | BinaryOp::Mul
-            | BinaryOp::And
-            | BinaryOp::Or
-            | BinaryOp::Xor
-            | BinaryOp::Shl
-            | BinaryOp::Shr
-            | BinaryOp::Sar,
-            rhs,
-        ) => {
-            dependency(*lhs);
-            dependency(*rhs);
-        }
-        SIRInstruction::Unary(
-            _,
-            UnaryOp::Ident
-            | UnaryOp::ToTwoState
-            | UnaryOp::Minus
-            | UnaryOp::Or
-            | UnaryOp::Xor
-            | UnaryOp::PopCount,
-            source,
-        ) => {
-            dependency(*source);
-        }
-        SIRInstruction::Concat(_, sources) => {
-            for &source in sources {
-                dependency(source);
-            }
-        }
-        SIRInstruction::Slice(_, source, _, _) => dependency(*source),
-        // Equal exact-zero arms make an unknown four-state condition
-        // irrelevant too, so the condition is not a dependency.
-        SIRInstruction::Mux(_, _, then_value, else_value) => {
-            dependency(*then_value);
-            dependency(*else_value);
-        }
-        _ => return None,
-    }
-    Some(count)
-}
-
-/// Prove exact all-zero values reachable from selected roots without
-/// materializing their potentially enormous bit representation or building a
-/// reverse-use graph for unrelated SIR. The explicit stack also avoids host
-/// recursion on deep expression chains.
-pub(crate) fn collect_exact_zero_registers<A>(
-    eu: &ExecutionUnit<A>,
-    roots: impl IntoIterator<Item = RegisterId>,
-) -> HashSet<RegisterId> {
-    let mut definitions = HashMap::<RegisterId, Option<&SIRInstruction<A>>>::default();
-    for block in eu.blocks.values() {
-        for instruction in &block.instructions {
-            if let Some(dst) = instruction.defined_register() {
-                if let Some(definition) = definitions.get_mut(&dst) {
-                    *definition = None;
-                } else {
-                    definitions.insert(dst, Some(instruction));
-                }
-            }
-        }
-    }
-
-    let mut result = HashMap::<RegisterId, bool>::default();
-    let mut visiting = HashSet::default();
-    for root in roots {
-        let mut work = vec![(root, false)];
-        while let Some((register, expanded)) = work.pop() {
-            if result.contains_key(&register) {
-                visiting.remove(&register);
-                continue;
-            }
-            let Some(instruction) = definitions.get(&register).copied().flatten() else {
-                result.insert(register, false);
-                visiting.remove(&register);
-                continue;
-            };
-            if expanded {
-                let mut all_zero = true;
-                let count = visit_exact_zero_dependencies(instruction, |dependency| {
-                    all_zero &= result.get(&dependency) == Some(&true);
-                });
-                result.insert(register, count.is_some_and(|count| count != 0) && all_zero);
-                visiting.remove(&register);
-                continue;
-            }
-            if let SIRInstruction::Imm(_, value) = instruction {
-                result.insert(register, value.payload.is_zero() && value.mask.is_zero());
-                continue;
-            }
-            if !visiting.insert(register) {
-                result.insert(register, false);
-                continue;
-            }
-            let mut count = 0usize;
-            work.push((register, true));
-            if visit_exact_zero_dependencies(instruction, |dependency| {
-                count += 1;
-                if !result.contains_key(&dependency) {
-                    work.push((dependency, false));
-                }
-            })
-            .is_none()
-                || count == 0
-            {
-                work.pop();
-                result.insert(register, false);
-                visiting.remove(&register);
-            }
-        }
-    }
-    result
-        .into_iter()
-        .filter_map(|(register, zero)| zero.then_some(register))
-        .collect()
-}
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 pub enum GlueAddrBase<V> {
     Parent(V),
